+++ R/get_captions.R
@@ -1,28 +1,28 @@
 #' Get Captions of a Video
 #' 
-#' @param part  Caption resource requested. Required. Comma separated list of one or more of the 
-#' following: \code{id, snippet}. e.g., "id, snippet", "id" Default: \code{snippet}.  
+#' A few of Youtube videos have caption tracks available from an older 
+#' YouTube API. If that caption track is available, this function returns that,
+#' Or it returns caption track specified by id resource. Check \code{\link{list_caption_tracks}} for more
+#' information.
+#' 
 #' @param video_id ID of the video whose captions are requested. Required. No default.
 #' @param lang  Language of the caption; required; default is english ("en")
-#' @param id    comma-separated list of IDs that identify the caption resources that should be retrieved; optional; string
+#' @param id    String. id of the caption track that is being retrieved
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return captions for the video from one of the first track
+#' @return String. 
 #' 
 #' @export
-#' 
-#' @references \url{https://developers.google.com/youtube/v3/docs/captions/list}
-#' 
+#'  
 #' @examples
 #' \dontrun{
 #' get_captions(video_id="yJXTXN4xrI8")
+#' get_captions(id="y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
 #' }
 
-get_captions <- function (part="snippet", video_id=NULL, lang="en", id=NULL, ...) {
+get_captions <- function (video_id=NULL, lang="en", id = NULL, ...) {
 
 	if (is.null(video_id)) stop("Must specify a video ID")
-
-	yt_check_token()
 
 	# Try getting captions directly
 	req <- GET(paste0("http://video.google.com/timedtext?lang=", lang, "&v=", video_id))
@@ -31,16 +31,17 @@ get_captions <- function (part="snippet", video_id=NULL, lang="en", id=NULL, ...
 	# If not try other things
 	if (length(req)==0) {
 		
-		querylist = list(part=part, videoId = video_id, id = id)
-		req <- tuber_GET("captions", query = querylist, ...)
+		querylist = list(id = id)
+		raw_res <- tuber_GET("captions", query = querylist, ...)
 		
-		# Multiple caption tracks possible but for now harvest just the first
-		caption_id <- req$items[[1]]$id
+		if (length(raw_res$items) ==0) { 
+    		cat("No caption tracks available. Likely cause: Incorrect video ID. \n")
+    		return(list())
+    	}
 
-		caption <- tuber_GET("captions", query=list(part=part, videoId = video_id, id = caption_id), ...)
-
-		return(caption)
+		return(captions)
 	}
 
 	req
 }
+